refactor: change window to contorl
