+++ Login.xaml.cs
@@ -14,80 +14,6 @@ namespace Project
             InitializeComponent();
         }
 
-        private void LoginButton_Click(object sender, RoutedEventArgs e)
-        {
-            string loginId = LoginIdBox.Text;
-            string password = PasswordBox.Password;
-
-            if (string.IsNullOrWhiteSpace(loginId) || string.IsNullOrWhiteSpace(password))
-            {
-                LoginErrorText.Text = "사번과 비밀번호를 모두 입력해주세요.";
-                LoginErrorText.Visibility = Visibility.Visible;
-                return;
-            }
-
-
-            bool loginSuccess = (loginId == "12345678" && password == "password123");
-
-            if (loginSuccess)
-            {
-                LoginErrorText.Visibility = Visibility.Collapsed;
-
-
-                System.Windows.MessageBox.Show("로그인 성공!");
-
-
-                Sign_up main = new Sign_up();
-                main.Show();
-                this.Close();
-            }
-            else
-            {
-                LoginErrorText.Text = "사번 또는 비밀번호가 잘못되었습니다.";
-                LoginErrorText.Visibility = Visibility.Visible;
-            }
-        }
-
-
-        private void PasswordBox_PasswordChanged(object sender, RoutedEventArgs e)
-        {
-            PasswordPlaceholder.Visibility = string.IsNullOrEmpty(PasswordBox.Password)
-                ? Visibility.Visible
-                : Visibility.Collapsed;
-        }
-
-
-        private void GoToSignUp_Click(object sender, MouseButtonEventArgs e)
-        {
-            Sign_up signUp = new Sign_up();
-            signUp.Show();
-            this.Close();
-        }
-        private void LoginIdBox_GotFocus(object sender, RoutedEventArgs e)
-        {
-            LoginIdPlaceholder.Visibility = Visibility.Collapsed;
-        }
-
-        private void LoginIdBox_LostFocus(object sender, RoutedEventArgs e)
-        {
-            if (string.IsNullOrWhiteSpace(LoginIdBox.Text))
-            {
-                LoginIdPlaceholder.Visibility = Visibility.Visible;
-            }
-        }
-        private void PasswordBox_GotFocus(object sender, RoutedEventArgs e)
-        {
-            PasswordPlaceholder.Visibility = Visibility.Collapsed;
-        }
-
-        private void PasswordBox_LostFocus(object sender, RoutedEventArgs e)
-        {
-            if (string.IsNullOrWhiteSpace(PasswordBox.Password))
-            {
-                PasswordPlaceholder.Visibility = Visibility.Visible;
-            }
-        }
-
 
         void insertQuery()
         {
