+++ Model/Auth.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Project
+namespace user_client.Model
 {
     internal class Auth
     {
