+++ View/SignInControl.xaml
@@ -1,0 +1,84 @@
+﻿<UserControl x:Class="user_client.View.SignInControl"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:user_client.View"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800">
+    <Grid>
+        <StackPanel Width="250" VerticalAlignment="Center" HorizontalAlignment="Center" Background="White" Margin="0,20">
+
+            <Grid Margin="0,0,0,10">
+                <TextBox x:Name="LoginIdBox"
+     Height="35"
+     FontSize="14"
+     VerticalContentAlignment="Center"
+     GotFocus="LoginIdBox_GotFocus"
+     LostFocus="LoginIdBox_LostFocus"
+     />
+
+
+                <TextBlock x:Name="LoginIdPlaceholder"
+       Text="사번을 입력하세요"
+       Foreground="#AAAAAA"
+       FontSize="14"
+       Margin="5,0,0,0"
+       VerticalAlignment="Center"
+       IsHitTestVisible="False"
+       Visibility="Visible"/>
+            </Grid>
+
+
+
+            <Grid Margin="0,0,0,20">
+                <PasswordBox x:Name="PasswordBox"
+                     Height="35"    
+                     FontSize="14"
+                     VerticalContentAlignment="Center"
+                     GotFocus="PasswordBox_GotFocus"
+                     LostFocus="PasswordBox_LostFocus"/>
+
+                <TextBlock x:Name="PasswordPlaceholder"
+                     Text="비밀번호를 입력하세요"
+                     Foreground="#AAAAAA"
+                     FontSize="14"
+                     Margin="5,0,0,0"
+                     VerticalAlignment="Center"
+                     IsHitTestVisible="False"
+                     Visibility="Visible"/>
+
+            </Grid>
+
+
+
+
+            <TextBlock x:Name="LoginErrorText"
+       Text="사번 또는 비밀번호가 잘못되었습니다"
+       Foreground="Red"
+       FontSize="12"
+       Visibility="Collapsed"
+       Margin="0,0,0,10"
+       HorizontalAlignment="Center" />
+
+
+            <Button Content="로그인"
+    Height="40"
+    FontWeight="Bold"
+    Background="#DDD"
+    BorderBrush="Gray"
+    Margin="0,0,0,10"
+    Click="LoginButton_Click"/>
+
+
+            <TextBlock Text="신규 입사자라면? 회원가입 하러 가기"
+       Foreground="#AAAAAA"
+       FontSize="11"
+       HorizontalAlignment="Center"
+       Cursor="Hand"
+       MouseDown="GoToSignUp_Click"/>
+
+        </StackPanel>
+
+    </Grid>
+</UserControl>
