+++ user-client.csproj.user
@@ -23,5 +23,11 @@
     <Compile Update="View\PostListControl.xaml.cs">
       <SubType>Code</SubType>
     </Compile>
+    <Compile Update="View\SignInControl.xaml.cs">
+      <SubType>Code</SubType>
+    </Compile>
+    <Compile Update="View\SignUpControl.xaml.cs">
+      <SubType>Code</SubType>
+    </Compile>
   </ItemGroup>
 </Project>
